requirements.txt
pillow==10.4.0
PyMuPDF==1.24.9
PyMuPDFb==1.24.9
